first pass at externalizing the ontology
the config file now defines the default ontology terms and there
is an onto service which provides access to them in their full namedspaced
and prefixed forms.  This is still a compromise because ideally the ontology
would be defined per collection and per item type in the template files for the
collections and items (i.e. the equivalent of the cite textinventory).  This is
too much work to do as the code is currently structured though.