ontology update
first pass at configuring the deployment ontology for image collections in Perseids.
There are some remaining terms that still need to be corrected --- these are mainly the
app specific fields, for which, if we use a custom ontology, I'd like the namespace
to be something different, but I'm not sure what. For now, the official fields of the cite
collection items should be mostly what we want. Need now to verify this by uploading some
items from an existing collection.